clustering api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
umap-learn==0.5.4
hdbscan==0.8.33
requests==2.31.0
python-multipart==0.0.6